added generic connect and ping health strategy
