+++ lib/health-strategies/generic/index.js
@@ -1,0 +1,1 @@
+module.exports.ConnectThenPingStrategy = require('./connect-then-ping')
