+++ lib/health-strategies/http/strategy.js
@@ -45,4 +45,4 @@ class HttpStrategy extends HealthStrategy {
   cleanup () {}
 }
 
-module.exports = HttpStrategy
+module.exports = (adapter, opts) => new HttpStrategy(adapter, opts)
