+++ lib/health-strategies/providers/strategy-factory-provider.js
@@ -16,6 +16,6 @@ module.exports = (Strategy, AdapterFactory) => {
      * @param {Object} [opts] - configiuration for an adapter
      * @return {Strategy}
      */
-    return (adapter = Adapter(), opts) => new Strategy(adapter, opts)
+    return (adapter = Adapter(), opts) => Strategy(adapter, opts)
   }
 }
